refactor: move sections to components + rename header and footer with uppercases
